MapView useEffect for dynamic marker addition - for dev, fetches all Places from DB on page load - for dev, uses lat lon, before PostGIS impl.
