+++ SportMap.Server/Program.cs
@@ -18,15 +18,6 @@ builder.Services.AddProblemDetails();
 builder.Services.AddOpenApi();
 builder.Services.AddControllers();
 builder.Services.AddDataProtection();
-builder.Services.AddCors(options =>
-{
-    options.AddPolicy("AllowAll", policy =>
-    {
-        policy.AllowAnyOrigin()
-              .AllowAnyMethod()
-              .AllowAnyHeader();
-    });
-});
 builder.Services.AddCookiePolicy(options =>
 {
     options.MinimumSameSitePolicy = builder.Environment.IsDevelopment()
@@ -91,7 +82,6 @@ if (app.Environment.IsDevelopment())
 }
 
 app.UseOutputCache();
-app.UseCors("AllowAll");
 app.MapDefaultEndpoints();
 app.UseCookiePolicy();
 app.UseAuthentication();
