+++ frontend/src/types/place.ts
@@ -1,0 +1,22 @@
+export interface PlaceDto {
+  id: string;
+  name: string;
+  description: string;
+  placeTypeId: string;
+  placeType?: PlaceTypeDto;
+  latitude: number;
+  longitude: number;
+  address?: string;
+  imageId?: string;
+  creatorId: string;
+  creatorName: string;
+  createdAt: string;
+  updatedAt?: string;
+  status: string;
+}
+
+export interface PlaceTypeDto {
+  id: string;
+  name: string;
+  description: string;
+}
